Remove about card
