--- src/components/organisms/index.ts
@@ -1,1 +1,0 @@
-export * from './AboutCard';
